requirements.txt
atlassian-python-api==3.41.3
requests==2.31.0
flask==3.0.0